+++ resources/views/components/pie-labels.blade.php
@@ -1,0 +1,5 @@
+@props(['series' => [], 'labels' => [], 'legend' => true])
+
+<x-chart::base component="pie-labels" :option="\WireCharts\Support\Option::pieLabels([
+    'series' => $series, 'labels' => $labels, 'legend' => $legend,
+])" {{ $attributes }} />